inDanger method progress
The inDanger piece method has been theoretically completed. It however
only works if used on black (AI) pieces which from a design standpoint
may be a bit flawed but for our purposes is fine so far. I've made it
so it differentiates between Kings and Pieces to recognized danger by
more than just proximity.