AI Methods
More strategizing and working of the AI methods have been done this
week. Credit will be given in commented code. I received very helpful
advice from Zach and Wyatt in regard to the functionality of the
program. The AI now has an ArrayList of actors it will use to further
decide what pieces to actually move.