Comments and Integration
Since problems regarding the global variable to manage the turn system
did not work, we moved to an idea suggested by Wyatt Smith. The idea is
to, instead of having a boolean keep track of turns, have the player
move methods input through the interface execute World.step() and
increment the grid forward automatically to move the AI pieces. This
idea worked well with his earlier suggested idea of having the AI run
its code as an actor on the grid. This allowed for other methods like
getGrid() to function as well as conveniently allow compatibility with
this new World.step() idea. Unfortunately there are still some errors
in the integration of AI and game that prevent us from really testing
the program as a whole.